Add files via upload
In agriculture, there is a need for a technology that is more easily understood, implemented, and used by the farmers. Equipment that requires less human effort and time with less cost of implementation is much required for success in the agricultural industry. Hence, we designed a Seed sowing robot that can be controlled with the help of IoT and it helps in the sowing of seeds in the desired position hence assisting the farmers in saving time and money. This machine performs the operations like plowing, sowing, and leveling which are used for small-scale farming. Seed sowing is one of the main processes of farming activities. It requires a substantial amount of human effort and also time-consuming. This project aims to design and fabricate a seed sowing robot for the mentioned task. This robot requires less maintenance and is portable.